+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The "ARCHIVE" column header to "REPOSITORY" for accuracy and consistency.
 ### Fixed
 - Minor copy/paste error in snap badge in README.md.
 - CVE links/URLs use https instead of http.